fix fill_triangle bug issue #33
This fix for issue #33 replaces the for loop at lines 283 to 291 with a while loop so that the y variable is properly incremented between the loops to fill the top and bottom parts of the triangle. This avoids screen row at y0 being drawn twice, and corrects the shape of the triangle. The changes to lines 278 to 282 (as numbered after the fix) initiate the two loops in a way that addresses issue #22 where a triangle with a flat bottom edge was not drawn correctly.

This fix has been implemented in the derived repository  [framebuf2](https://github.com/peter-l5/framebuf2) which applies the algorithm here and has been  tested with the following code. (Note that `f=True` parameter draws a filled triangle in this implementation.)
```
    for z in range(3,123+1,20):
        for x in range(10,120+1,20):
            for y in range(0,128):
                display.fill(0)
                display.triangle(x,y,30,z,90,63,c=1,f=True)
                display.show()
```